+++ src/lib/index.ts
@@ -4,3 +4,4 @@ export * from './bars';
 export * from './points';
 export * from './legend';
 export * from './filters';
+export * from './chart-window';